gen: use int64 for some fields
With int32, some values could not be encoded like volume sizes which are expressed in bytes.
An int32 could only code for 4Gib max.
This patch fix yaml file in order to generate int64 integers instead of int32 ones.
This will be fixed in official outscale yaml.

+++ osc/model_create_snapshot_request.go
@@ -18,7 +18,7 @@ type CreateSnapshotRequest struct {
 	// The pre-signed URL of the snapshot you want to import from the OSU bucket.
 	FileLocation string `json:"FileLocation,omitempty"`
 	// The size of the snapshot created in your account, in bytes. This size must be exactly the same as the source snapshot one.
-	SnapshotSize int32 `json:"SnapshotSize,omitempty"`
+	SnapshotSize int64 `json:"SnapshotSize,omitempty"`
 	// The name of the source Region, which must be the same as the Region of your account.
 	SourceRegionName string `json:"SourceRegionName,omitempty"`
 	// The ID of the snapshot you want to copy.
